requirements.txt
streamlit>=1.35
httpx>=0.27
python-dotenv>=1.0
pdfplumber>=0.11
python-docx>=1.1
Pillow>=11.0
PyMuPDF>=1.24
